add boundary.spec.tsx
